Add option to have one subshell per target + merge settings
