--- packages/apputils-extension/schema/kernels-subshells.json
@@ -1,13 +1,0 @@
-{
-  "title": "Kernel Subshells",
-  "description": "Settings for the kernel subshells usage.",
-  "type": "object",
-  "properties": {
-    "commsOverSubshells": {
-      "type": "boolean",
-      "title": "Kernel Comms over subshells",
-      "description": "Whether comm messages should be sent to kernel subshells, if the kernel supports it.",
-      "default": false
-    }
-  }
-}
