+++ packages/apputils-extension/src/statusbarplugin.ts
@@ -13,13 +13,10 @@ import {
   ISessionContext,
   ISessionContextDialogs,
   KernelStatus,
-  RunningSessions,
   SessionContextDialogs
 } from '@jupyterlab/apputils';
-import { ISettingRegistry } from '@jupyterlab/settingregistry';
 import { IStatusBar } from '@jupyterlab/statusbar';
 import { ITranslator, nullTranslator } from '@jupyterlab/translation';
-import { IDisposable } from '@lumino/disposable';
 import { Title, Widget } from '@lumino/widgets';
 import { KeyboardEvent } from 'react';
 
@@ -129,82 +126,5 @@ export const kernelStatus: JupyterFrontEndPlugin<IKernelStatusModel> = {
     });
 
     return { addSessionProvider };
-  }
-};
-
-/*
- * A plugin providing running terminals and sessions information
- * to the status bar.
- */
-export const runningSessionsStatus: JupyterFrontEndPlugin<void> = {
-  id: '@jupyterlab/apputils-extension:running-sessions-status',
-  description: 'Add the running sessions and terminals status bar item.',
-  autoStart: true,
-  requires: [IStatusBar, ITranslator],
-  optional: [ISettingRegistry],
-  activate: (
-    app: JupyterFrontEnd,
-    statusBar: IStatusBar,
-    translator: ITranslator,
-    settingRegistry: ISettingRegistry | null
-  ) => {
-    const createStatusItem = () => {
-      const item = new RunningSessions({
-        onClick: () => app.shell.activateById('jp-running-sessions'),
-        onKeyDown: (event: KeyboardEvent<HTMLImageElement>) => {
-          if (
-            event.key === 'Enter' ||
-            event.key === 'Spacebar' ||
-            event.key === ' '
-          ) {
-            event.preventDefault();
-            event.stopPropagation();
-            app.shell.activateById('jp-running-sessions');
-          }
-        },
-        serviceManager: app.serviceManager,
-        translator
-      });
-
-      item.model.sessions = Array.from(
-        app.serviceManager.sessions.running()
-      ).length;
-      item.model.terminals = Array.from(
-        app.serviceManager.terminals.running()
-      ).length;
-
-      return item;
-    };
-
-    const registerItem = () => {
-      const item = createStatusItem();
-      return statusBar.registerStatusItem(runningSessionsStatus.id, {
-        item,
-        align: 'left',
-        rank: 0
-      });
-    };
-
-    if (settingRegistry) {
-      let disposable: IDisposable;
-      const onSettingsUpdated = (
-        settings: ISettingRegistry.ISettings
-      ): void => {
-        const showStatusBarItem = settings.get('showStatusBarItem')
-          .composite as boolean;
-
-        disposable?.dispose();
-        if (showStatusBarItem) {
-          disposable = registerItem();
-        }
-      };
-
-      void settingRegistry.load(runningSessionsStatus.id).then(settings => {
-        onSettingsUpdated(settings);
-        settings.changed.connect(onSettingsUpdated);
-      });
-    } else {
-      registerItem();
-    }
   }
 };
