+++ packages/services/src/kernel/index.ts
@@ -2,10 +2,11 @@
 // Distributed under the terms of the Modified BSD License.
 
 // Namespace some of our modules for convenience and backwards compatibility.
+import { CommsOverSubshells } from './comm';
 import * as Kernel from './kernel';
 import * as KernelMessage from './messages';
 import * as KernelAPI from './restapi';
 import { KernelConnection } from './default';
 
 export * from './manager';
-export { Kernel, KernelMessage, KernelAPI, KernelConnection };
+export { Kernel, KernelMessage, KernelAPI, KernelConnection, CommsOverSubshells };
