+++ packages/services/src/kernel/kernel.ts
@@ -7,7 +7,7 @@ import { IDisposable, IObservableDisposable } from '@lumino/disposable';
 
 import { ISignal } from '@lumino/signaling';
 
-import { ServerConnection } from '..';
+import { CommsOverSubshells, ServerConnection } from '..';
 
 import * as KernelMessage from './messages';
 
@@ -122,7 +122,7 @@ export interface IKernelConnection extends IObservableDisposable {
    * If enabled, we'll create one subshell per-comm, this may lead to issue
    * if many comms are open, so it's disabled by default.
    */
-  commsOverSubshells?: boolean;
+  commsOverSubshells?: CommsOverSubshells;
 
   /**
    * Whether the kernel connection has pending input.
@@ -678,7 +678,7 @@ export namespace IKernelConnection {
      * If enabled, we'll create one subshell per-comm, this may lead to issue
      * if many comms are open, so it's disabled by default.
      */
-    commsOverSubshells?: boolean;
+    commsOverSubshells?: CommsOverSubshells;
 
     /**
      * The unique identifier for the kernel client.
@@ -744,7 +744,7 @@ export interface IManager extends IBaseManager {
    * If enabled, we'll create one subshell per-comm, this may lead to issue
    * if many comms are open, so it's disabled by default.
    */
-  commsOverSubshells: boolean;
+  commsOverSubshells: CommsOverSubshells;
 
   /**
    * Force a refresh of the running kernels.
